+++ src/wizards/audience/views/setup/content-gifting.js
@@ -5,11 +5,7 @@
  */
 import { __ } from '@wordpress/i18n';
 import {
-	ExternalLink,
 	RangeControl,
-	SelectControl,
-	TextControl,
-	Button,
 	BaseControl,
 	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
 	__experimentalHeading as Heading,
@@ -18,10 +14,11 @@ import {
 	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
 	__experimentalToggleGroupControlOption as ToggleGroupControlOption,
 } from '@wordpress/components';
-import { ActionCard, Grid, Notice } from '../../../../../packages/components/src';
+import { ActionCard, Button, Grid, Notice, SelectControl, TextControl } from '../../../../../packages/components/src';
 
 export default function ContentGifting( { config, setConfig, updateConfig, noBorder = false } ) {
 	const giftingErrors = Object.values( newspackAudience?.content_gifting?.can_use_gifting?.errors || {} ).flat();
+	const availableProducts = newspackAudience?.available_products || [];
 	const hasMetering = newspackAudience?.content_gifting?.has_metering;
 
 	return (
@@ -38,7 +35,7 @@ export default function ContentGifting( { config, setConfig, updateConfig, noBor
 			{ config.content_gifting?.enabled && (
 				<>
 					{ giftingErrors.length > 0 && <Notice noticeText={ giftingErrors.join( ', ' ) } isError /> }
-					<Grid columns={ 2 } rowGap={ 32 }>
+					<Grid columns={ 2 } rowGap={ 16 }>
 						<Heading level={ 4 } style={ { gridColumn: '1 / -1' } }>
 							{ __( 'General Settings', 'newspack-plugin' ) }
 						</Heading>
@@ -86,7 +83,10 @@ export default function ContentGifting( { config, setConfig, updateConfig, noBor
 							__next40pxDefaultSize
 						/>
 					</Grid>
-					<Grid columns={ 2 } rowGap={ 32 }>
+					<Grid columns={ 1 }>
+						<hr />
+					</Grid>
+					<Grid columns={ 2 } rowGap={ 16 }>
 						<Heading level={ 4 } style={ { gridColumn: '1 / -1' } }>
 							{ __( 'Recipient Banner', 'newspack-plugin' ) }
 						</Heading>
@@ -104,23 +104,8 @@ export default function ContentGifting( { config, setConfig, updateConfig, noBor
 							onChange={ value => setConfig( { ...config, content_gifting: { ...config.content_gifting, button_label: value } } ) }
 							__next40pxDefaultSize
 						/>
-						<TextControl
-							label={ __( 'Subscribe button URL', 'newspack-plugin' ) }
-							help={
-								<>
-									{ __(
-										'URL for the subscribe button in the banner. If not provided, the primary subscription tier product will be used with modal checkout.',
-										'newspack-plugin'
-									) }{ ' ' }
-									<ExternalLink href="/wp-admin/admin.php?page=newspack-audience-subscriptions">
-										{ __( 'Configure the primary subscription product', 'newspack-plugin' ) }
-									</ExternalLink>
-								</>
-							}
-							value={ config.content_gifting.cta_url }
-							onChange={ value => setConfig( { ...config, content_gifting: { ...config.content_gifting, cta_url: value } } ) }
-							__next40pxDefaultSize
-						/>
+					</Grid>
+					<Grid columns={ 3 } rowGap={ 16 }>
 						<ToggleGroupControl
 							label={ __( 'Style', 'newspack-plugin' ) }
 							value={ config.content_gifting.style || 'light' }
@@ -131,9 +116,45 @@ export default function ContentGifting( { config, setConfig, updateConfig, noBor
 							<ToggleGroupControlOption label={ __( 'Light', 'newspack-plugin' ) } value="light" />
 							<ToggleGroupControlOption label={ __( 'Dark', 'newspack-plugin' ) } value="dark" />
 						</ToggleGroupControl>
+						<ToggleGroupControl
+							label={ __( 'Subscribe button action', 'newspack-plugin' ) }
+							help={ __(
+								'Whether the subscribe button should start a product checkout or redirect to a landing page.',
+								'newspack-plugin'
+							) }
+							value={ config.content_gifting.cta_type || 'product' }
+							onChange={ value => setConfig( { ...config, content_gifting: { ...config.content_gifting, cta_type: value } } ) }
+							isBlock
+							__next40pxDefaultSize
+						>
+							<ToggleGroupControlOption label={ __( 'Product', 'newspack-plugin' ) } value="product" />
+							<ToggleGroupControlOption label={ __( 'Landing page', 'newspack-plugin' ) } value="url" />
+						</ToggleGroupControl>
+						{ config.content_gifting.cta_type === 'product' && (
+							<SelectControl
+								label={ __( 'Subscribe button product', 'newspack-plugin' ) }
+								help={ __( 'Product linked to the subscribe button.', 'newspack-plugin' ) }
+								options={ [ { label: __( 'Select a product', 'newspack-plugin' ), value: 0, disabled: true }, ...availableProducts ] }
+								value={ config.content_gifting.cta_product_id }
+								suggestions={ availableProducts.map( o => o.label ) }
+								onChange={ value =>
+									setConfig( { ...config, content_gifting: { ...config.content_gifting, cta_product_id: value } } )
+								}
+								__next40pxDefaultSize
+							/>
+						) }
+						{ config.content_gifting.cta_type === 'url' && (
+							<TextControl
+								label={ __( 'Subscribe button URL', 'newspack-plugin' ) }
+								help={ __( 'URL for the landing page to redirect to.', 'newspack-plugin' ) }
+								value={ config.content_gifting.cta_url }
+								onChange={ value => setConfig( { ...config, content_gifting: { ...config.content_gifting, cta_url: value } } ) }
+								__next40pxDefaultSize
+							/>
+						) }
 						<div style={ { gridColumn: '1 / -1' } }>
 							<BaseControl id="newspack-content-gifting-cta-preview" label={ __( 'Preview', 'newspack-plugin' ) }>
-								<div className="newspack-content-gifting__cta-preview" inert>
+								<div className="newspack-content-gifting__cta-preview" inert="true">
 									<div className="newspack-ui">
 										<div
 											className={ `banner newspack-content-gifting__cta is-style-${ config.content_gifting.style || 'light' }` }
@@ -153,15 +174,18 @@ export default function ContentGifting( { config, setConfig, updateConfig, noBor
 														) }
 													</div>
 												</div>
-												<button
-													className={ `newspack-ui__button newspack-ui__button--x-small ${
-														( config.content_gifting.style || 'light' ) === 'dark'
-															? 'newspack-ui__button--primary-light'
-															: 'newspack-ui__button--accent'
-													}` }
-												>
-													{ config.content_gifting.button_label || __( 'Subscribe now', 'newspack-plugin' ) }
-												</button>
+												{ ( ( config.content_gifting.cta_type === 'product' && config.content_gifting.cta_product_id ) ||
+													( config.content_gifting.cta_type === 'url' && config.content_gifting.cta_url ) ) && (
+													<button
+														className={ `newspack-ui__button newspack-ui__button--x-small ${
+															( config.content_gifting.style || 'light' ) === 'dark'
+																? 'newspack-ui__button--primary-light'
+																: 'newspack-ui__button--accent'
+														}` }
+													>
+														{ config.content_gifting.button_label || __( 'Subscribe now', 'newspack-plugin' ) }
+													</button>
+												) }
 											</div>
 										</div>
 									</div>